refactor: Lotto -> Lottery 이름 변경
