+++ README.md
@@ -4,7 +4,7 @@
 
 - [X] 로또는 중복되지 않는 1~45 사이의 6개의 숫자를 가진다
 - [X] 유효한 로또를 자동으로 생성한다
-- [ ] 구입 금액에 맞춰 로또 여러 장을 자동으로 생성한다
+- [X] 구입 금액에 맞춰 로또 여러 장을 자동으로 생성한다
 - [ ] 로또가 당첨 번호와 몇 개가 일치하는지 알아낸다
 - [ ] 구입한 로또들 중 n(3~6)개 일치하는 것들이 몇 개 있는지 알아낸다
 - [ ] 총 수익을 계산한다
