+++ src/main/java/camp/nextstep/edu/lottery/LotteryGeneratr.java
@@ -1,23 +1,23 @@
-package camp.nextstep.edu.lotto;
+package camp.nextstep.edu.lottery;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class LottoGenerator {
+public class LotteryGeneratr {
 
     public static final int LOTTO_NUMBER_COUNT = 6;
     public static final int LOTTO_NUMBER_START = 1;
     public static final int LOTTO_NUMBER_END = 45;
 
-    static Lotto create() {
+    static Lottery create() {
         List<Integer> numbers = IntStream
             .rangeClosed(LOTTO_NUMBER_START, LOTTO_NUMBER_END)
             .boxed()
             .collect(Collectors.toList());
         Collections.shuffle(numbers);
 
-        return new Lotto(numbers.subList(0, LOTTO_NUMBER_COUNT));
+        return new Lottery(numbers.subList(0, LOTTO_NUMBER_COUNT));
     }
 }
